libavformat feature - alternative library for media info
