+++ src/services/mod.rs
@@ -30,6 +30,7 @@ pub mod search;
 mod subs;
 pub mod transcode;
 mod types;
+pub mod audio_meta;
 
 const APP_STATIC_FILES_CACHE_AGE: u32 = 30 * 24 * 3600;
 const FOLDER_INFO_FILES_CACHE_AGE: u32 = 24 * 3600;
